+++ packages/telemetry-service/tests/telemetryService.spec.ts
@@ -1,6 +1,6 @@
 // Copyright 2024 IOTA Stiftung.
 // SPDX-License-Identifier: Apache-2.0.
-import { TelemetryConnectorFactory, type ITelemetryConnector } from "@gtsc/telemetry-models";
+import { TelemetryConnectorFactory, type ITelemetryConnector } from "@twin.org/telemetry-models";
 import { TelemetryService } from "../src/telemetryService";
 
 describe("TelemetryService", () => {